add Makefile and split train and tests

+++ game/action.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 @dataclass
 class Action:
     thrust: int
-    angle: float  # will be converted to X, Y after
+    angle: int
 
     def __str__(self):
         return f"{self.thrust},{self.angle}"
